# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: CasaRoca System AI · API
  version: "1.0.0"
  description: |
    API del sistema eclesial de Casa Sobre la Roca: una central y 36 sedes.

    ESTE ARCHIVO SE GENERA. No lo edite a mano: corra `scripts/generar-openapi.js`.
    La verificacion de entrega compara lo generado con lo versionado.

    Autenticacion: token de portador (Bearer). El token dice que sesion DICE ser;
    quien decide es la base, que en cada peticion responde si esa sesion sigue viva.
    Por eso cerrar una sesion surte efecto en el instante.
servers:
  - url: https://api.casaroca.org
    description: produccion
  - url: https://api.staging.casaroca.org
    description: pruebas
  - url: http://127.0.0.1:3000
    description: desarrollo
components:
  securitySchemes:
    portador:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: boolean }
        mensaje: { type: string, description: "En castellano y con la accion siguiente. Nunca filtra la estructura interna." }
        peticionId: { type: string, description: "Identificador de traza. El usuario lo lee por telefono al soporte." }
security:
  - portador: []
paths:
...
  /api/v1/formacion/cohortes/:id/inscribir:
    post:
      summary: "Inscribir a alguien"
      description: "Si la cohorte esta llena responde 409 y lo dice (cuenta igual que la base). En una cohorte con valor exige `estadoPago` (pendiente, pagado, parcial o exonerado)."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/formacion/cohortes/:id:
    get:
      summary: "La ficha de la cohorte"
      description: "Quien esta inscrito, como va y quien pago."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/formacion/cohortes:
    get:
      summary: "Las cohortes"
      description: "Son de cada sede: cada una abre su grupo cuando puede y con el docente que tiene. Avisa de las que pasaron el cupo."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    post:
      summary: "Abrir una cohorte"
      description: "Curso, sede, codigo, modalidad y fechas."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/formacion/cursos:
    get:
      summary: "Los cursos"
      description: "Con su programa, su semestre y su prerequisito."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/formacion/inscripciones/:id/calificar:
    post:
      summary: "Calificar"
      description: "Nota final y estado. Sin nota CONSERVA la anterior (antes la borraba); borrarla exige `borrarNota: true`. El pago no se cambia aqui: tiene su propia ruta."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/formacion/inscripciones/:id/pago:
    post:
      summary: "Registrar el pago de una inscripcion"
      description: "Cambia el estado del pago (no_aplica, pendiente, pagado, exonerado, parcial) con una `referencia` opcional que va a la auditoria. No existia: una inscripcion pendiente no podia pasar nunca a pagada."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/formacion/programas:
    get:
      summary: "Los programas de la red"
      description: "El catalogo lo define la central una vez y lo usan las 36 sedes."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...